Fix Python version in the coverage Action
This change fixes the version of Python the coverage Action uses.

With the version of Python that tox is looking for not available, tox doesn't return non-zero, so the check passes even though it reports `coverage: skipped because could not find python interpreter with spec(s): python3.8`. This fixes that issue.

Co-authored-by: Andy Chosak <andy.chosak@cfpb.gov>