Change how breadcrumbs appear
Prior to this change, breadcrumbs from our books example would follow this pattern:

J. R. R. Tolkein -> The Lord of the Rings -> Volumes

Where the author name would link to the full list of authors, the book name would link to the list of books under the author, and the book’s volumes would exist under the model name itself.

This change makes that same set of breadcrumbs appear as:

Authors -> J. R. R. Tolkien -> The Lord of the Rings

Where “Authors” links to the full list of authors, the author name links to the authors books, and the book title links to the book’s volumes.

In converting the breadcrumbs to Wagtail 4, this seemed more natural.