Remove upper version pin
This change removes our upper version pin of Wagtail, enabling (slightly broken) support for Wagtail 6.

With Wagtail 6 there are some additional changes to the breadcrumb and listing templates that we should bring in. The current listing works okay, with the exception of the buttons, but breadcrumbs are entirely broken. The breadcrumbs are a primary means of navigating the model tree, so that'll need to be fixed before this app officially supports Wagtail 6.x.

This change should make this app non-blocking for Wagtail 6.x though.